requirements.txt
pdfplumber
tqdm
python-dotenv

openai
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic

faiss-cpu